backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
Pillow>=10.0.0
python-multipart>=0.0.6
